cleanup moonsniff code
