+++ examples/moonsniff/post-processing.lua
@@ -1,19 +1,8 @@
 --- Demonstrates the basic usage of moonsniff in order to determine device induced latencies
 
-local lm        = require "libmoon"
-local device    = require "device"
-local memory    = require "memory"
-local ts        = require "timestamping"
-local hist      = require "histogram"
-local timer     = require "timer"
 local log       = require "log"
-local stats     = require "stats"
-local barrier   = require "barrier"
-local ms	= require "moonsniff-io"
-local bit	= require "bit"
 local dpdk	= require "dpdk"
 local pcap	= require "pcap"
-local hmap	= require "hmap"
 local profile 	= require "jit.p"
 
 local ffi    = require "ffi"
@@ -21,30 +10,10 @@ local C = ffi.C
 
 -- default values when no cli options are specified
 local INPUT_PATH = "latencies.csv"
-local INPUT_MODE = C.ms_text
-local BITMASK = 0x0FFFFFFF
 local TIME_THRESH = -50 	-- negative timevalues smaller than this value are not allowed
 
 local MODE_MSCAP, MODE_PCAP = 0, 1
 local MODE = MODE_MSCAP
-
--- pointers and ctypes
-local CHAR_P = ffi.typeof("char *")
-local INT64_T = ffi.typeof("int64_t")
-local UINT64_P = ffi.typeof("uint64_t *")
-local UINT8_T = ffi.typeof("uint8_t")
-local UINT8_P = ffi.typeof("uint8_t*")
-
-local free = C.rte_pktmbuf_free_export
-local band = bit.band
-
-local pktmatch = nil
-local scratchpad = nil
-local SCR_SIZE = 16 -- size of the scratchpad in bytes, must always be multiple of 8 for hash to work
-local mempool = nil
-local mempool2 = nil
-local next_mem = 0
-local SIP_KEY = ffi.new("uint64_t[2]", {1, 2})
 
 -- skip the initialization of DPDK, as it is not needed for this script
 dpdk.skipInit()
@@ -57,18 +26,10 @@ function configure(parser)
 	parser:option("-n --nrbuckets", "Size of a bucket for the resulting histogram"):args(1):convert(tonumber):default(1)
 	parser:flag("-b --binary", "Read a file which was generated by moonsniff with the binary flag set")
 	parser:flag("-d --debug", "Create additional debug information")
+	parser:flag("-p --profile", "Profile the application. May decrease the overall performance")
         return parser:parse()
 end
 
-ffi.cdef[[
-	void* malloc(size_t);
-	void free(void*);
-
-	uint32_t ms_hash(void*);
-	uint32_t ms_get_identifier(void*);
-
-	uint64_t SipHashC(const uint64_t* key, const char* bytes, const uint64_t size);
-]]
 
 function master(args)
 	if args.input then INPUT_PATH = args.input end
@@ -80,6 +41,8 @@ function master(args)
 
 	elseif string.match(args.input, ".*%.mscap") then
 		MODE = MODE_MSCAP
+	else
+		log:err("Input with unknown file extension.\nCan only process .pcap or .mscap files.")
 	end
 
 	print(MODE)
@@ -117,20 +80,43 @@ function master(args)
 		end
 	end
 
-	if MODE == MODE_PCAP or MODE == MODE_MSCAP then
 
-		print(PRE)
-		print(POST)
+	print(PRE)
+	print(POST)
 
-		-- use new tbb matching mode
-		if MODE == MODE_PCAP then
-			local tbb = require "tbbmatch"
-			tbb.match(PRE, POST, args)
-			return
-		elseif MODE == MODE_MSCAP then
-			local arr = require "arrmatch"
-			arr.match(PRE, POST, args)
-			return
-		end
+
+	-- retrieve additional information about input
+	local file = assert(io.open(PRE, "r"))
+        local size = fsize(file)
+        file:close()
+        file = assert(io.open(POST, "r"))
+        size = size + fsize(file)
+        file:close()
+        log:info("File size: " .. size / 1e9 .. " [GB]")
+        local nClock = os.clock()
+        profile.start("-fl", "somefile.txt")
+
+	-- run correct matching
+	if MODE == MODE_PCAP then
+		local tbb = require "tbbmatch"
+		tbb.match(PRE, POST, args)
+	elseif MODE == MODE_MSCAP then
+		local arr = require "arrmatch"
+		arr.match(PRE, POST, args)
 	end
+
+	-- finish operations
+	profile.stop()
+
+        local elapsed = os.clock() - nClock
+        log:info("Elapsed time core: " .. elapsed .. " [sec]")
+        log:info("Processing speed: " .. (size / 1e6) / elapsed .. " [MB/s]")
+end
+
+--- Compute the size of a file
+function fsize(file)
+        local current = file:seek()
+        local size = file:seek("end")
+        file:seek("set", current)
+        return size
 end
